Combine page break specs into a single spec file
